feat: move content-watcher type & constant defs to shared location
