+++ apps/content-publishing-api/src/metadata.ts
@@ -1,5 +1,9 @@
 /* eslint-disable */
 export default async () => {
-    const t = {};
-    return { "@nestjs/swagger": { "models": [] } };
+    const t = {
+        ["../../../libs/types/src/dtos/content-publishing/activity.dto"]: await import("../../../libs/types/src/dtos/content-publishing/activity.dto"),
+        ["../../../libs/types/src/dtos/content-publishing/announcement.dto"]: await import("../../../libs/types/src/dtos/content-publishing/announcement.dto"),
+        ["../../../libs/types/src/dtos/account/wallet.login.request.dto"]: await import("../../../libs/types/src/dtos/account/wallet.login.request.dto")
+    };
+    return { "@nestjs/swagger": { "models": [[import("../../../libs/types/src/dtos/content-publishing/common.dto"), { "DsnpUserIdParam": { userDsnpId: { required: true, type: () => String } }, "AnnouncementResponseDto": { referenceId: { required: true, type: () => String } }, "UploadResponseDto": { assetIds: { required: true, type: () => [String] } }, "FilesUploadDto": { files: { required: true, type: () => [Object] } } }], [import("../../../libs/types/src/dtos/content-publishing/activity.dto"), { "LocationDto": { name: { required: true, type: () => String, minLength: 1 }, accuracy: { required: false, type: () => Number, minimum: 0, maximum: 100 }, altitude: { required: false, type: () => Number }, latitude: { required: false, type: () => Number }, longitude: { required: false, type: () => Number }, radius: { required: false, type: () => Number, minimum: 0 }, units: { required: false, enum: t["../../../libs/types/src/dtos/content-publishing/activity.dto"].UnitTypeDto } }, "AssetReferenceDto": { referenceId: { required: true, type: () => String, minLength: 1 }, height: { required: false, type: () => Number, minimum: 1 }, width: { required: false, type: () => Number, minimum: 1 }, duration: { required: false, type: () => String, pattern: "DURATION_REGEX" } }, "TagDto": { type: { required: true, enum: t["../../../libs/types/src/dtos/content-publishing/activity.dto"].TagTypeDto }, name: { required: false, type: () => String, minLength: 1 }, mentionedId: { required: false, type: () => String } }, "AssetDto": { references: { required: false, type: () => [t["../../../libs/types/src/dtos/content-publishing/activity.dto"].AssetReferenceDto] }, name: { required: false, type: () => String, minLength: 1 }, href: { required: false, type: () => String, minLength: 1 } }, "BaseActivityDto": { name: { required: false, type: () => String }, tag: { required: false, type: () => [t["../../../libs/types/src/dtos/content-publishing/activity.dto"].TagDto] }, location: { required: false, type: () => t["../../../libs/types/src/dtos/content-publishing/activity.dto"].LocationDto } }, "NoteActivityDto": { content: { required: true, type: () => String, minLength: 1 }, published: { required: true, type: () => String }, assets: { required: false, type: () => [t["../../../libs/types/src/dtos/content-publishing/activity.dto"].AssetDto] } }, "ProfileActivityDto": { icon: { required: false, type: () => [t["../../../libs/types/src/dtos/content-publishing/activity.dto"].AssetReferenceDto] }, summary: { required: false, type: () => String }, published: { required: false, type: () => String } } }], [import("../../../libs/types/src/dtos/content-publishing/announcement.dto"), { "BroadcastDto": { content: { required: true, type: () => t["../../../libs/types/src/dtos/content-publishing/activity.dto"].NoteActivityDto } }, "ReplyDto": { inReplyTo: { required: true, type: () => String }, content: { required: true, type: () => t["../../../libs/types/src/dtos/content-publishing/activity.dto"].NoteActivityDto } }, "TombstoneDto": { targetContentHash: { required: true, type: () => String }, targetAnnouncementType: { required: true, enum: t["../../../libs/types/src/dtos/content-publishing/announcement.dto"].ModifiableAnnouncementTypeDto } }, "UpdateDto": { targetContentHash: { required: true, type: () => String }, targetAnnouncementType: { required: true, enum: t["../../../libs/types/src/dtos/content-publishing/announcement.dto"].ModifiableAnnouncementTypeDto }, content: { required: true, type: () => t["../../../libs/types/src/dtos/content-publishing/activity.dto"].NoteActivityDto } }, "ReactionDto": { emoji: { required: true, type: () => String, minLength: 1, pattern: "DSNP_EMOJI_REGEX" }, apply: { required: true, type: () => Number, minimum: 0, maximum: 255 }, inReplyTo: { required: true, type: () => String } }, "ProfileDto": { profile: { required: true, type: () => t["../../../libs/types/src/dtos/content-publishing/activity.dto"].ProfileActivityDto } } }], [import("../../../libs/types/src/dtos/account/handles.request.dto"), { "HandleRequestDto": { accountId: { required: true, type: () => String }, proof: { required: true, type: () => String } }, "ChangeHandlePayloadRequest": { encodedPayload: { required: true, type: () => String } } }], [import("../../../libs/types/src/dtos/account/wallet.login.request.dto"), { "ErrorResponseDto": { message: { required: true, type: () => String } }, "SiwsPayloadDto": { message: { required: true, type: () => String }, signature: { required: true, type: () => String } }, "SignInResponseDto": { siwsPayload: { required: false, type: () => t["../../../libs/types/src/dtos/account/wallet.login.request.dto"].SiwsPayloadDto }, error: { required: false, type: () => t["../../../libs/types/src/dtos/account/wallet.login.request.dto"].ErrorResponseDto } }, "EncodedExtrinsicDto": { pallet: { required: true, type: () => String }, extrinsicName: { required: true, type: () => String }, encodedExtrinsic: { required: true, type: () => String } }, "SignUpResponseDto": { extrinsics: { required: false, type: () => [t["../../../libs/types/src/dtos/account/wallet.login.request.dto"].EncodedExtrinsicDto] }, error: { required: false, type: () => t["../../../libs/types/src/dtos/account/wallet.login.request.dto"].ErrorResponseDto } }, "WalletLoginRequestDto": { signIn: { required: true, type: () => t["../../../libs/types/src/dtos/account/wallet.login.request.dto"].SignInResponseDto }, signUp: { required: true, type: () => t["../../../libs/types/src/dtos/account/wallet.login.request.dto"].SignUpResponseDto } } }], [import("../../../libs/types/src/dtos/account/keys.request.dto"), { "KeysRequestDto": { msaOwnerAddress: { required: true, type: () => String }, msaOwnerSignature: { required: true, type: () => String }, newKeyOwnerSignature: { required: true, type: () => String } } }]], "controllers": [[import("./controllers/health.controller"), { "HealthController": { "healthz": {}, "livez": {}, "readyz": {} } }]] } };
 };
