--- libs/content-watcher-lib/src/index.ts
@@ -1,6 +1,0 @@
-export * from './dtos/announcement.dto';
-export * from './dtos/activity.dto';
-export * from './dtos/common.dto';
-export * from './dtos/validation.dto';
-export * from './dtos/content-search-request.dto';
-export * from './queues/queue-constants';
