+++ libs/content-watcher-lib/src/pubsub/announcers/tombstone.ts
@@ -1,10 +1,10 @@
 import { Processor } from '@nestjs/bullmq';
 import { Injectable } from '@nestjs/common';
 import { Job } from 'bullmq';
-import * as QueueConstants from '../../queues/queue-constants';
+import * as QueueConstants from '#types/constants/content-watcher.queue.constants';
 import { BaseConsumer } from '../../utils/base-consumer';
 import { PubSubService } from '../pubsub.service';
-import { AnnouncementResponse } from '../../types/content-announcement';
+import { AnnouncementResponse } from '#types/content-announcement';
 
 @Injectable()
 @Processor(QueueConstants.TOMBSTONE_QUEUE_NAME, { concurrency: 2 })
