+++ libs/types/src/constants/content-watcher.constants.ts
@@ -1,8 +1,4 @@
 /**
- * Number of seconds to create a block in Frequency chain
- */
-export const SECONDS_PER_BLOCK = 12;
-/**
  * Name of timeout event used for in memory scheduler
  */
 export const CAPACITY_EPOCH_TIMEOUT_NAME = 'capacity-epoch-timeout';
