+++ libs/types/src/content-announcement/index.ts
