+++ libs/types/src/content-announcement/types.gen.ts
