+++ libs/types/src/dtos/content-watcher/activity.dto.ts
