+++ libs/types/src/dtos/content-watcher/announcement.dto.ts
