+++ libs/types/src/dtos/content-watcher/chain.watch.dto.ts
