+++ libs/types/src/dtos/content-watcher/content-search-request.dto.ts
