+++ libs/types/src/dtos/content-watcher/index.ts
@@ -1,0 +1,7 @@
+export * from './activity.dto';
+export * from './announcement.dto';
+export * from './chain.watch.dto';
+export * from './common.dto';
+export * from './content-search-request.dto';
+export * from './subscription.webhook.dto';
+export * from './validation.dto';
