+++ libs/types/src/dtos/content-watcher/subscription.webhook.dto.ts
