+++ libs/types/src/dtos/content-watcher/validation.dto.ts
