+++ libs/types/src/interfaces/content-watcher/announcement-subscription.interface.ts
