+++ libs/types/src/interfaces/content-watcher/index.ts
@@ -1,0 +1,4 @@
+export * from './announcement-subscription.interface';
+export * from './ipfs.job.interface';
+export * from './message_response_with_schema_id';
+export * from './scan-reset.interface';
