+++ libs/types/src/interfaces/content-watcher/ipfs.job.interface.ts
