+++ libs/types/src/interfaces/content-watcher/message_response_with_schema_id.ts
