+++ libs/types/src/interfaces/content-watcher/scan-reset.interface.ts
