+++ package.json
@@ -59,7 +59,7 @@
     "generate:openapi:graph": "nest start graph-api --entryFile build-openapi",
     "generate:openapi": "npm run generate:openapi:account ; npm run generate:openapi:content-publishing ; npm run generate:openapi:content-watcher ; npm run generate:openapi:graph",
     "generate:types:graph": "npx openapi-client-axios-typegen openapi-specs/graph-webhooks.openapi.yaml > libs/graph-lib/src/types/webhook-types.d.ts ; prettier --write libs/graph-lib/src/types/webhook-types.d.ts",
-    "generate:types:content-watcher": "npx @hey-api/openapi-ts -i openapi-specs/content-announcement.openapi.json -o libs/content-watcher-lib/src/types/content-announcement ; prettier --write libs/content-watcher-lib/src/types/content-announcement",
+    "generate:types:content-watcher": "npx @hey-api/openapi-ts -i openapi-specs/content-announcement.openapi.json -o libs/types/src/content-announcement ; prettier --write libs/types/src/content-announcement",
     "generate:types": "npm run generate:types:graph ; npm run generate:types:content-watcher",
     "generate:swagger-ui:account": "npx --yes @redocly/cli build-docs openapi-specs/account.openapi.json --output=./docs/account/index.html",
     "generate:swagger-ui:content-publishing": "npx --yes @redocly/cli build-docs openapi-specs/content-publishing.openapi.json --output=./docs/content-publishing/index.html",
